+++ mitra-activitypub/src/lib.rs
@@ -2,3 +2,4 @@ pub mod agent;
 pub mod constants;
 pub mod fetch;
 pub mod http_client;
+pub mod utils;